+++ src/main/java/sonia/scm/jenkins/GlobalJenkinsConfiguration.java
@@ -24,32 +24,36 @@
 
 package sonia.scm.jenkins;
 
-//~--- non-JDK imports --------------------------------------------------------
-
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import sonia.scm.Validateable;
 import sonia.scm.util.Util;
 
-//~--- JDK imports ------------------------------------------------------------
-
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
 
-/**
- *
- * @author Sebastian Sdorra
- */
 @XmlAccessorType(XmlAccessType.FIELD)
 @XmlRootElement(name = "jenkins-config")
 @Getter
 @Setter
 @NoArgsConstructor
-public class GlobalJenkinsConfiguration implements Validateable
-{
+public class GlobalJenkinsConfiguration implements Validateable {
+  @XmlElement(name = "disable-repository-configuration")
+  private boolean disableRepositoryConfiguration = false;
+
+  @XmlElement(name = "disable-mercurial-trigger")
+  private boolean disableMercurialTrigger = false;
+
+  @XmlElement(name = "disable-git-trigger")
+  private boolean disableGitTrigger = false;
+
+  @XmlElement(name = "disable-event-trigger")
+  private boolean disableEventTrigger = false;
+
+  private String url;
 
   public GlobalJenkinsConfiguration(String url, boolean disableGitTrigger, boolean disableMercurialTrigger, boolean disableRepositoryConfiguration) {
     this.url = url;
@@ -59,25 +63,8 @@ public class GlobalJenkinsConfiguration implements Validateable
   }
 
   @Override
-  public boolean isValid()
-  {
+  public boolean isValid() {
     return Util.isNotEmpty(url);
   }
 
-  //~--- fields ---------------------------------------------------------------
-
-  /** Field description */
-  @XmlElement(name = "disable-repository-configuration")
-  private boolean disableRepositoryConfiguration = false;
-
-  /** Field description */
-  @XmlElement(name = "disable-mercurial-trigger")
-  private boolean disableMercurialTrigger = false;
-
-  /** Field description */
-  @XmlElement(name = "disable-git-trigger")
-  private boolean disableGitTrigger = false;
-
-  /** Field description */
-  private String url;
 }