+++ src/main/java/sonia/scm/jenkins/JenkinsEventRelay.java
@@ -1,0 +1,111 @@
+/*
+ * MIT License
+ *
+ * Copyright (c) 2020-present Cloudogu GmbH and Contributors
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+package sonia.scm.jenkins;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import de.otto.edison.hal.HalRepresentation;
+import de.otto.edison.hal.Link;
+import de.otto.edison.hal.Links;
+import lombok.Getter;
+import lombok.Setter;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import sonia.scm.config.ScmConfiguration;
+import sonia.scm.net.ahc.AdvancedHttpClient;
+import sonia.scm.repository.Repository;
+import sonia.scm.repository.api.RepositoryServiceFactory;
+import sonia.scm.repository.api.ScmProtocol;
+
+import java.io.IOException;
+import java.util.List;
+import java.util.stream.Collectors;
+
+public abstract class JenkinsEventRelay {
+
+  private static final Logger logger = LoggerFactory.getLogger(JenkinsEventRelay.class);
+  public static final String EVENT_ENDPOINT = "scm-manager-hook/notify";
+  private final ObjectMapper mapper = new ObjectMapper();
+
+  private final ScmConfiguration configuration;
+
+  protected final JenkinsContext jenkinsContext;
+  protected final RepositoryServiceFactory repositoryServiceFactory;
+  protected final AdvancedHttpClient httpClient;
+
+  public JenkinsEventRelay(ScmConfiguration configuration, JenkinsContext jenkinsContext, RepositoryServiceFactory repositoryServiceFactory, AdvancedHttpClient httpClient) {
+    this.configuration = configuration;
+    this.jenkinsContext = jenkinsContext;
+    this.repositoryServiceFactory = repositoryServiceFactory;
+    this.httpClient = httpClient;
+  }
+
+  protected final void send(Repository repository, JenkinsEventDto eventDto) {
+    if (!jenkinsContext.getConfiguration().isDisableEventTrigger()) {
+      eventDto.setServer(configuration.getBaseUrl());
+      eventDto.setNamespace(repository.getNamespace());
+      eventDto.setName(repository.getName());
+      eventDto.setType(repository.getType());
+
+      try {
+        String json = mapper.writer().writeValueAsString(eventDto);
+
+        httpClient.post(createEventHookUrl(repository))
+          .formContent().field("json", json).build()
+          .request();
+      } catch (IOException e) {
+        if (logger.isWarnEnabled()) {
+          logger.warn("Failed to relay event to Jenkins server");
+        }
+      }
+    }
+  }
+
+  private String createEventHookUrl(Repository repository) {
+    String url;
+
+    if (jenkinsContext.getConfiguration().isDisableRepositoryConfiguration()) {
+      url = jenkinsContext.getConfiguration().getUrl();
+    } else {
+      url = jenkinsContext.getConfiguration(repository).getUrl();
+    }
+    if (!url.endsWith("/")) {
+      url = url + "/";
+    }
+    return url + EVENT_ENDPOINT;
+  }
+
+  @Getter
+  @Setter
+  protected static class JenkinsEventDto extends HalRepresentation {
+    private String namespace;
+    private String name;
+    private String type;
+    private String server;
+
+    JenkinsEventDto(List<ScmProtocol> protocols) {
+      super(new Links.Builder().array(protocols.stream().map(protocol -> Link.link(protocol.getType(), protocol.getUrl())).collect(Collectors.toList())).build());
+    }
+  }
+}